Build ProjectsPage and do big refactor of DataTable
DataTable is now a reusable component used in both PeoplePage and
DataTable. PeoplePage had to be refactored too, to use this new
DataTable. Most significant refactor is: DataTable requires many
properties (props), specially specifications of its columns.

Add usage of ProjectsPage to main component, and link to it on NavBar.