Fix TextFilter functionality on ProjectsPage
And refactor also PeoplePage TextFilter so that it doesn't get mixed up
with the ProjectsPage TextFilter. The virtual-dom needs a key for these
to know that they are completely distinct (rather than being just a
modification). Also refactored code in models of both PeoplePage and
ProjectsPage to update the search filter value in the state object.